moved svg to assets
